refactor: moved move logic out of TradeBasicConfirmDetails
